requirements.txt
Flask==3.0.2
paramiko==3.4.0
python-dotenv==1.0.1 